+++ lib/Cathedral/Builder/EntityAbstractBuilder.php
@@ -31,8 +31,14 @@ use Zend\Filter\Null;
  */
 class EntityAbstractBuilder extends BuilderAbstract implements BuilderInterface {
 	
+	/**
+	 * @var string
+	 */
 	protected $type = self::TYPE_ENTITYABSTRACT;
 	
+	/* (non-PHPdoc)
+	 * @see \Cathedral\Builder\BuilderAbstract::setupFile()
+	 */
 	protected function setupFile() {
 		$this->_file->setNamespace($this->getNames()->namespace_entity);
 	
@@ -52,6 +58,11 @@ class EntityAbstractBuilder extends BuilderAbstract implements BuilderInterface 
 		return $prepend.str_replace(' ','',ucwords(str_replace('_',' ',$property)));
 	}
 	
+	/**
+	 * Create getter & setter methods for properties
+	 * 
+	 * @param string $property
+	 */
 	protected function addGetterSetter($property) {
 		$properyName = $this->parseMethodName($property, '');
 		$getter = "get{$properyName}";
@@ -64,7 +75,14 @@ class EntityAbstractBuilder extends BuilderAbstract implements BuilderInterface 
 return \$this->{$property};
 MBODY;
 		$method->setBody($body);
+		$method->setDocBlock(DocBlockGenerator::fromArray([
+			'shortDescription' => "Get the {$property} property",
+			'tags' => [
+			new ReturnTag(['datatype' => $this->getNames()->properties[$property]['type']])
+			]]));
 		$this->_class->addMethodFromGenerator($method);
+		
+		//===============================================
 		
 		// METHOD:setPropperty
 		$parameterSetter = new ParameterGenerator();
@@ -75,9 +93,20 @@ MBODY;
 \$this->{$property} = \${$property};
 MBODY;
 		$method->setBody($body);
+		$method->setDocBlock(DocBlockGenerator::fromArray([
+			'shortDescription' => "Set the {$property} property",
+			'tags' => [
+			new ParamTag($property, ['datatype' => $this->getNames()->properties[$property]['type']])
+			]]));
 		$this->_class->addMethodFromGenerator($method);
 	}
 	
+	/**
+	 * Create method to return related Parent entity
+	 *  linked to foreign key stored in this coloumn
+	 * 
+	 * @param string $columnName
+	 */
 	protected function addRelationParent($columnName) {
 		$table = substr($columnName, 3);
 		$parent = new NameManager($this->getNames()->namespace, $table);
@@ -97,6 +126,12 @@ MBODY;
 		$this->_class->addMethodFromGenerator($method);
 	}
 	
+	/**
+	 * Create method to return related children entities
+	 *  this primary key found in table
+	 *  
+	 * @param string $tableName
+	 */
 	protected function addRelationChild($tableName) {
 		$child = new NameManager($this->getNames()->namespace, $tableName);
 		
@@ -116,6 +151,9 @@ MBODY;
 		$this->_class->addMethodFromGenerator($method);
 	}
 	
+	/* (non-PHPdoc)
+	 * @see \Cathedral\Builder\BuilderAbstract::setupClass()
+	 */
 	protected function setupClass() {
 		$this->_class->setName($this->getNames()->entityAbstractName);
 		$this->_class->setImplementedInterfaces(['RowGatewayInterface']);
@@ -132,6 +170,10 @@ MBODY;
 				$property->setDefaultValue($value['default']);
 			}
 			
+			$property->setDocBlock(DocBlockGenerator::fromArray([
+				'tags' => [[
+					'name' => 'var',
+					'description' => $value['type']]]]));
 			$this->_class->addPropertyFromGenerator($property);
 		}
 		
@@ -149,6 +191,9 @@ MBODY;
 		$this->_file->setClass($this->_class);
 	}
 	
+	/* (non-PHPdoc)
+	 * @see \Cathedral\Builder\BuilderAbstract::setupMethods()
+	 */
 	protected function setupMethods() {
 		//PARAMETERS
 		$parameterPrimary = new ParameterGenerator();