+++ lib/Cathedral/Builder/Exception/PermissionException.php
@@ -22,6 +22,9 @@ namespace Cathedral\Builder\Exception;
  */
 class PermissionException extends \RuntimeException implements ExceptionInterface {
 	
+	/* (non-PHPdoc)
+	 * @see \Cathedral\Builder\Exception\ExceptionInterface::getCallingClass()
+	 */
 	protected function getCallingClass() {
 		$d = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS, 3);
 		$class = $d[2]["class"];
@@ -32,12 +35,20 @@ class PermissionException extends \RuntimeException implements ExceptionInterfac
 		return $class;
 	}
 	
+	/* (non-PHPdoc)
+	 * @see \Cathedral\Builder\Exception\ExceptionInterface::callingFunction()
+	*/
 	private function callingFunction() {
 		$d = debug_backtrace();
 		$function = $d[2]["function"];
 		return $function;
 	}
 	
+	/**
+	 * Create exception with message
+	 *
+	 * @param string $message
+	 */
 	public function __construct($message) {
 		$class = $this->getCallingClass();
 		$function = $this->callingFunction();